+++ test/backend/test_backend.py
@@ -4,10 +4,6 @@ from libqtile.backend import get_core
 from libqtile.utils import QtileError
 
 
-def test_get_core_x11(display):
-    get_core('x11', display).finalize()
-
-
 def test_get_core_bad():
     with pytest.raises(QtileError):
         get_core("NonBackend").finalize()
