+++ test/backend/x11/test_xcore.py
@@ -1,4 +1,11 @@
+import pytest
+
+from libqtile.backend import get_core
 from libqtile.backend.x11 import core
+
+
+def test_get_core_x11(display):
+    get_core('x11', display).finalize()
 
 
 def test_keys(display):
@@ -6,10 +13,11 @@ def test_keys(display):
     assert "shift" in core.get_modifiers()
 
 
-def test_no_two_qtiles(manager):
-    try:
-        core.Core(manager.display).finalize()
-    except core.ExistingWMException:
-        pass
-    else:
-        raise Exception("expected an error on multiple qtiles connecting")
+def test_no_two_qtiles(xmanager):
+    with pytest.raises(core.ExistingWMException):
+        core.Core(xmanager.display).finalize()
+
+
+def test_color_pixel(xmanager):
+    (success, e) = xmanager.c.eval("self.core.conn.color_pixel(\"ffffff\")")
+    assert success, e
