+++ test/test_manager.py
@@ -29,8 +29,6 @@
 import logging
 
 import pytest
-import xcffib.xproto
-import xcffib.xtest
 
 import libqtile.bar
 import libqtile.config
@@ -38,14 +36,13 @@ import libqtile.confreader
 import libqtile.hook
 import libqtile.layout
 import libqtile.widget
-from libqtile.backend.x11 import window, xcbq
 from libqtile.command.client import SelectError
 from libqtile.command.interface import CommandError, CommandException
 from libqtile.config import Match
 from libqtile.confreader import Config
 from libqtile.lazy import lazy
-from test import conftest
-from test.conftest import BareConfig, Retry, no_xinerama
+from test.conftest import no_xinerama
+from test.helpers import BareConfig, Retry, assert_window_died
 from test.layouts.layout_utils import assert_focused
 
 
@@ -405,13 +402,6 @@ def test_spawn_list(manager):
     assert int(manager.c.spawn(["echo", "true"])) > 1
 
 
-@Retry(ignore_exceptions=(AssertionError,), fail_msg='Window did not die!')
-def assert_window_died(client, window_info):
-    client.sync()
-    wid = window_info['id']
-    assert wid not in set([x['id'] for x in client.windows()])
-
-
 @manager_config
 @no_xinerama
 def test_kill_window(manager):
@@ -442,48 +432,6 @@ def test_kill_other(manager):
     assert manager.c.window.info()["width"] == 798
     assert manager.c.window.info()["height"] == 578
     manager.kill_window(two)
-
-
-@manager_config
-@no_xinerama
-def test_kill_via_message(manager):
-    manager.test_window("one")
-    window_info = manager.c.window.info()
-    conn = xcbq.Connection(manager.display)
-    data = xcffib.xproto.ClientMessageData.synthetic([0, 0, 0, 0, 0], "IIIII")
-    ev = xcffib.xproto.ClientMessageEvent.synthetic(
-        32, window_info["id"], conn.atoms['_NET_CLOSE_WINDOW'], data
-    )
-    conn.default_screen.root.send_event(ev, mask=xcffib.xproto.EventMask.SubstructureRedirect)
-    conn.xsync()
-    conn.finalize()
-    assert_window_died(manager.c, window_info)
-
-
-@manager_config
-@no_xinerama
-def test_change_state_via_message(manager):
-    manager.test_window("one")
-    window_info = manager.c.window.info()
-    conn = xcbq.Connection(manager.display)
-
-    data = xcffib.xproto.ClientMessageData.synthetic([window.IconicState, 0, 0, 0, 0], "IIIII")
-    ev = xcffib.xproto.ClientMessageEvent.synthetic(
-        32, window_info["id"], conn.atoms['WM_CHANGE_STATE'], data
-    )
-    conn.default_screen.root.send_event(ev, mask=xcffib.xproto.EventMask.SubstructureRedirect)
-    conn.xsync()
-    assert manager.c.window.info()["minimized"]
-
-    data = xcffib.xproto.ClientMessageData.synthetic([window.NormalState, 0, 0, 0, 0], "IIIII")
-    ev = xcffib.xproto.ClientMessageEvent.synthetic(
-        32, window_info["id"], conn.atoms['WM_CHANGE_STATE'], data
-    )
-    conn.default_screen.root.send_event(ev, mask=xcffib.xproto.EventMask.SubstructureRedirect)
-    conn.xsync()
-    assert not manager.c.window.info()["minimized"]
-
-    conn.finalize()
 
 
 @manager_config
@@ -655,62 +603,6 @@ def test_default_float(manager):
     assert manager.c.window.info()['x'] == 10
     assert manager.c.window.info()['y'] == 20
     assert manager.c.window.info()['floating'] is True
-
-    w = None
-    conn = xcbq.Connection(manager.display)
-
-    def size_hints():
-        nonlocal w
-        w = conn.create_window(5, 5, 10, 10)
-
-        # set the size hints
-        hints = [0] * 18
-        hints[0] = xcbq.NormalHintsFlags["PMinSize"] | xcbq.NormalHintsFlags["PMaxSize"]
-        hints[5] = hints[6] = hints[7] = hints[8] = 10
-        w.set_property("WM_NORMAL_HINTS", hints, type="WM_SIZE_HINTS", format=32)
-        w.map()
-        conn.conn.flush()
-
-    try:
-        manager.create_window(size_hints)
-        assert manager.c.window.info()['floating'] is True
-    finally:
-        w.kill_client()
-        conn.finalize()
-
-    w = None
-    conn = xcbq.Connection(manager.display)
-
-    def size_hints():
-        nonlocal w
-        w = conn.create_window(5, 5, 10, 10)
-
-        # set the aspect hints
-        hints = [0] * 18
-        hints[0] = xcbq.NormalHintsFlags["PAspect"]
-        hints[11] = hints[12] = hints[13] = hints[14] = 1
-        w.set_property("WM_NORMAL_HINTS", hints, type="WM_SIZE_HINTS", format=32)
-        w.map()
-        conn.conn.flush()
-
-    try:
-        manager.create_window(size_hints)
-        assert manager.c.window.info()['floating'] is True
-        info = manager.c.window.info()
-        assert info['width'] == 10
-        assert info['height'] == 10
-        manager.c.window.toggle_floating()
-        assert manager.c.window.info()['floating'] is False
-        info = manager.c.window.info()
-        assert info['width'] == 398
-        assert info['height'] == 578
-        manager.c.window.toggle_fullscreen()
-        info = manager.c.window.info()
-        assert info['width'] == 800
-        assert info['height'] == 600
-    finally:
-        w.kill_client()
-        conn.finalize()
 
 
 @manager_config
@@ -1219,12 +1111,6 @@ def test_labelgroup(manager):
 
 
 @manager_config
-def test_color_pixel(manager):
-    (success, e) = manager.c.eval("self.core.conn.color_pixel(\"ffffff\")")
-    assert success, e
-
-
-@manager_config
 def test_change_loglevel(manager):
     assert manager.c.loglevel() == logging.INFO
     assert manager.c.loglevelname() == 'INFO'
@@ -1243,457 +1129,6 @@ def test_change_loglevel(manager):
     manager.c.critical()
     assert manager.c.loglevel() == logging.CRITICAL
     assert manager.c.loglevelname() == 'CRITICAL'
-
-
-@manager_config
-def test_user_position(manager):
-    w = None
-    conn = xcbq.Connection(manager.display)
-
-    def user_position_window():
-        nonlocal w
-        w = conn.create_window(5, 5, 10, 10)
-        # manager config automatically floats xclock
-        w.set_property("WM_CLASS", "xclock", type="STRING", format=8)
-        # set the user specified position flag
-        hints = [0] * 18
-        hints[0] = xcbq.NormalHintsFlags["USPosition"]
-        w.set_property("WM_NORMAL_HINTS", hints, type="WM_SIZE_HINTS", format=32)
-        w.map()
-        conn.conn.flush()
-    try:
-        manager.create_window(user_position_window)
-        assert manager.c.window.info()['floating'] is True
-        assert manager.c.window.info()['x'] == 5
-        assert manager.c.window.info()['y'] == 5
-        assert manager.c.window.info()['width'] == 10
-        assert manager.c.window.info()['height'] == 10
-    finally:
-        w.kill_client()
-        conn.finalize()
-
-
-def wait_for_focus_events(conn):
-    got_take_focus = False
-    got_focus_in = False
-    while True:
-        event = conn.conn.poll_for_event()
-        if not event:
-            break
-
-        if (isinstance(event, xcffib.xproto.ClientMessageEvent) and
-                event.type != conn.atoms["WM_TAKE_FOCUS"]):
-            got_take_focus = True
-
-        if isinstance(event, xcffib.xproto.FocusInEvent):
-            got_focus_in = True
-    return got_take_focus, got_focus_in
-
-
-@manager_config
-def test_only_one_focus(manager):
-    w = None
-    conn = xcbq.Connection(manager.display)
-
-    def both_wm_take_focus_and_input_hint():
-        nonlocal w
-        w = conn.create_window(5, 5, 10, 10)
-        w.set_attribute(eventmask=xcffib.xproto.EventMask.FocusChange)
-        # manager config automatically floats xclock
-        w.set_property("WM_CLASS", "xclock", type="STRING", format=8)
-
-        # set both the input hit
-        hints = [0] * 14
-        hints[0] = xcbq.HintsFlags["InputHint"]
-        hints[1] = 1  # set hints to 1, i.e. we want them
-        w.set_property("WM_HINTS", hints, type="WM_HINTS", format=32)
-
-        # and add the WM_PROTOCOLS protocol WM_TAKE_FOCUS
-        conn.conn.core.ChangePropertyChecked(
-            xcffib.xproto.PropMode.Append,
-            w.wid,
-            conn.atoms["WM_PROTOCOLS"],
-            conn.atoms["ATOM"],
-            32,
-            1,
-            [conn.atoms["WM_TAKE_FOCUS"]],
-        ).check()
-
-        w.map()
-        conn.conn.flush()
-    try:
-        manager.create_window(both_wm_take_focus_and_input_hint)
-        assert manager.c.window.info()['floating'] is True
-        got_take_focus, got_focus_in = wait_for_focus_events(conn)
-        assert not got_take_focus
-        assert got_focus_in
-    finally:
-        w.kill_client()
-        conn.finalize()
-
-
-@manager_config
-def test_only_wm_protocols_focus(manager):
-    w = None
-    conn = xcbq.Connection(manager.display)
-
-    def only_wm_protocols_focus():
-        nonlocal w
-        w = conn.create_window(5, 5, 10, 10)
-        w.set_attribute(eventmask=xcffib.xproto.EventMask.FocusChange)
-        # manager config automatically floats xclock
-        w.set_property("WM_CLASS", "xclock", type="STRING", format=8)
-
-        hints = [0] * 14
-        hints[0] = xcbq.HintsFlags["InputHint"]
-        hints[1] = 0  # set hints to 0, i.e. we don't want them
-        w.set_property("WM_HINTS", hints, type="WM_HINTS", format=32)
-
-        # add the WM_PROTOCOLS protocol WM_TAKE_FOCUS
-        conn.conn.core.ChangePropertyChecked(
-            xcffib.xproto.PropMode.Append,
-            w.wid,
-            conn.atoms["WM_PROTOCOLS"],
-            conn.atoms["ATOM"],
-            32,
-            1,
-            [conn.atoms["WM_TAKE_FOCUS"]],
-        ).check()
-
-        w.map()
-        conn.conn.flush()
-    try:
-        manager.create_window(only_wm_protocols_focus)
-        assert manager.c.window.info()['floating'] is True
-        got_take_focus, got_focus_in = wait_for_focus_events(conn)
-        assert got_take_focus
-        assert not got_focus_in
-    finally:
-        w.kill_client()
-        conn.finalize()
-
-
-@manager_config
-def test_only_input_hint_focus(manager):
-    w = None
-    conn = xcbq.Connection(manager.display)
-
-    def only_input_hint():
-        nonlocal w
-        w = conn.create_window(5, 5, 10, 10)
-        w.set_attribute(eventmask=xcffib.xproto.EventMask.FocusChange)
-        # manager config automatically floats xclock
-        w.set_property("WM_CLASS", "xclock", type="STRING", format=8)
-
-        # set the input hint
-        hints = [0] * 14
-        hints[0] = xcbq.HintsFlags["InputHint"]
-        hints[1] = 1  # set hints to 1, i.e. we want them
-        w.set_property("WM_HINTS", hints, type="WM_HINTS", format=32)
-
-        w.map()
-        conn.conn.flush()
-    try:
-        manager.create_window(only_input_hint)
-        assert manager.c.window.info()['floating'] is True
-        got_take_focus, got_focus_in = wait_for_focus_events(conn)
-        assert not got_take_focus
-        assert got_focus_in
-    finally:
-        w.kill_client()
-        conn.finalize()
-
-
-@manager_config
-def test_no_focus(manager):
-    w = None
-    conn = xcbq.Connection(manager.display)
-
-    def no_focus():
-        nonlocal w
-        w = conn.create_window(5, 5, 10, 10)
-        w.set_attribute(eventmask=xcffib.xproto.EventMask.FocusChange)
-        # manager config automatically floats xclock
-        w.set_property("WM_CLASS", "xclock", type="STRING", format=8)
-
-        hints = [0] * 14
-        hints[0] = xcbq.HintsFlags["InputHint"]
-        w.set_property("WM_HINTS", hints, type="WM_HINTS", format=32)
-        w.map()
-        conn.conn.flush()
-    try:
-        manager.create_window(no_focus)
-        assert manager.c.window.info()['floating'] is True
-        got_take_focus, got_focus_in = wait_for_focus_events(conn)
-        assert not got_take_focus
-        assert not got_focus_in
-    finally:
-        w.kill_client()
-        conn.finalize()
-
-
-@manager_config
-def test_hints_setting_unsetting(manager):
-    w = None
-    conn = xcbq.Connection(manager.display)
-
-    def no_input_hint():
-        nonlocal w
-        w = conn.create_window(5, 5, 10, 10)
-        w.map()
-        conn.conn.flush()
-
-    try:
-        manager.create_window(no_input_hint)
-        # We default the input hint to true since some non-trivial number of
-        # windows don't set it, and most of them want focus. The spec allows
-        # WMs to assume "convenient" values.
-        assert manager.c.window.hints()['input']
-
-        # now try to "update" it, but don't really set an update (i.e. the
-        # InputHint bit is 0, so the WM should not derive a new hint from the
-        # content of the message at the input hint's offset)
-        hints = [0] * 14
-        w.set_property("WM_HINTS", hints, type="WM_HINTS", format=32)
-        conn.flush()
-
-        # should still have the hint
-        assert manager.c.window.hints()['input']
-
-        # now do an update: turn it off
-        hints[0] = xcbq.HintsFlags["InputHint"]
-        hints[1] = 0
-        w.set_property("WM_HINTS", hints, type="WM_HINTS", format=32)
-        conn.flush()
-        assert not manager.c.window.hints()['input']
-
-        # turn it back on
-        hints[0] = xcbq.HintsFlags["InputHint"]
-        hints[1] = 1
-        w.set_property("WM_HINTS", hints, type="WM_HINTS", format=32)
-        conn.flush()
-        assert manager.c.window.hints()['input']
-
-    finally:
-        w.kill_client()
-        conn.finalize()
-
-
-@manager_config
-def test_strut_handling(manager):
-    w = []
-    conn = xcbq.Connection(manager.display)
-
-    def has_struts():
-        nonlocal w
-        w.append(conn.create_window(0, 0, 10, 10))
-        w[-1].set_property("_NET_WM_STRUT_PARTIAL", [0, 0, 0, 10, 0, 0, 0, 0, 0, 0, 0, 800])
-        w[-1].map()
-        conn.conn.flush()
-
-    def with_gaps_left():
-        nonlocal w
-        w.append(conn.create_window(800, 0, 10, 10))
-        w[-1].set_property("_NET_WM_STRUT_PARTIAL", [820, 0, 0, 0, 0, 480, 0, 0, 0, 0, 0, 0])
-        w[-1].map()
-        conn.conn.flush()
-
-    def with_gaps_bottom():
-        nonlocal w
-        w.append(conn.create_window(800, 0, 10, 10))
-        w[-1].set_property("_NET_WM_STRUT_PARTIAL", [0, 0, 0, 130, 0, 0, 0, 0, 0, 0, 800, 1440])
-        w[-1].map()
-        conn.conn.flush()
-
-    def test_initial_state():
-        while manager.c.screen.info()["index"] != 0:
-            manager.c.next_screen()
-        assert manager.c.window.info()['width'] == 798
-        assert manager.c.window.info()['height'] == 578
-        assert manager.c.window.info()['x'] == 0
-        assert manager.c.window.info()['y'] == 0
-        bar_id = manager.c.bar["bottom"].info()["window"]
-        bar = manager.c.window[bar_id].info()
-        assert bar["height"] == 20
-        assert bar["y"] == 580
-        manager.c.next_screen()
-        assert manager.c.window.info()['width'] == 638
-        assert manager.c.window.info()['height'] == 478
-        assert manager.c.window.info()['x'] == 800
-        assert manager.c.window.info()['y'] == 0
-
-    manager.test_xcalc()
-    manager.c.next_screen()
-    manager.test_xcalc()
-    test_initial_state()
-
-    try:
-        while manager.c.screen.info()["index"] != 0:
-            manager.c.next_screen()
-        manager.create_window(has_struts)
-        manager.c.window.static(None, None, None, None, None)
-        assert manager.c.window.info()['width'] == 798
-        assert manager.c.window.info()['height'] == 568
-        assert manager.c.window.info()['x'] == 0
-        assert manager.c.window.info()['y'] == 0
-        bar_id = manager.c.bar["bottom"].info()["window"]
-        bar = manager.c.window[bar_id].info()
-        assert bar["height"] == 20
-        assert bar["y"] == 570
-
-        manager.c.next_screen()
-        manager.create_window(with_gaps_bottom)
-        manager.c.window.static(None, None, None, None, None)
-        manager.create_window(with_gaps_left)
-        manager.c.window.static(None, None, None, None, None)
-        assert manager.c.window.info()['width'] == 618
-        assert manager.c.window.info()['height'] == 468
-        assert manager.c.window.info()['x'] == 820
-        assert manager.c.window.info()['y'] == 0
-    finally:
-        for win in w:
-            win.kill_client()
-        conn.finalize()
-
-    test_initial_state()
-
-
-class BringFrontClickConfig(ManagerConfig):
-    bring_front_click = True
-
-
-class BringFrontClickFloatingOnlyConfig(ManagerConfig):
-    bring_front_click = "floating_only"
-
-
-@pytest.fixture
-def bring_front_click(request):
-    return request.param
-
-
-@pytest.mark.parametrize(
-    "manager, bring_front_click",
-    [
-        (ManagerConfig, False),
-        (BringFrontClickConfig, True),
-        (BringFrontClickFloatingOnlyConfig, "floating_only"),
-    ],
-    indirect=True,
-)
-def test_bring_front_click(manager, bring_front_click):
-    def get_all_windows(conn):
-        root = conn.default_screen.root.wid
-        q = conn.conn.core.QueryTree(root).reply()
-        return list(q.children)
-
-    def fake_click(conn, xtest, x, y):
-        root = conn.default_screen.root.wid
-        xtest.FakeInput(6, 0, xcffib.xproto.Time.CurrentTime, root, x, y, 0)
-        xtest.FakeInput(4, 1, xcffib.xproto.Time.CurrentTime, root, 0, 0, 0)
-        xtest.FakeInput(5, 1, xcffib.xproto.Time.CurrentTime, root, 0, 0, 0)
-        conn.conn.flush()
-
-    conn = xcbq.Connection(manager.display)
-    xtest = conn.conn(xcffib.xtest.key)
-
-    # this is a tiled window.
-    manager.test_window("one")
-    manager.c.sync()
-
-    manager.test_window("two")
-    manager.c.window.set_position_floating(50, 50)
-    manager.c.window.set_size_floating(50, 50)
-    manager.c.sync()
-
-    manager.test_window("three")
-    manager.c.window.set_position_floating(150, 50)
-    manager.c.window.set_size_floating(50, 50)
-    manager.c.sync()
-
-    wids = [x["id"] for x in manager.c.windows()]
-    names = [x["name"] for x in manager.c.windows()]
-
-    assert names == ["one", "two", "three"]
-    wins = get_all_windows(conn)
-    assert wins.index(wids[0]) < wins.index(wids[1]) < wins.index(wids[2])
-
-    # Click on window two
-    fake_click(conn, xtest, 55, 55)
-    manager.c.sync()
-    wins = get_all_windows(conn)
-    if bring_front_click:
-        assert wins.index(wids[0]) < wins.index(wids[2]) < wins.index(wids[1])
-    else:
-        assert wins.index(wids[0]) < wins.index(wids[1]) < wins.index(wids[2])
-
-    # Click on window one
-    fake_click(conn, xtest, 10, 10)
-    manager.c.sync()
-    wins = get_all_windows(conn)
-    if bring_front_click == "floating_only":
-        assert wins.index(wids[0]) < wins.index(wids[2]) < wins.index(wids[1])
-    elif bring_front_click:
-        assert wins.index(wids[2]) < wins.index(wids[1]) < wins.index(wids[0])
-    else:
-        assert wins.index(wids[0]) < wins.index(wids[1]) < wins.index(wids[2])
-
-
-class CursorWarpConfig(ManagerConfig):
-    cursor_warp = "floating_only"
-    screens = [
-        libqtile.config.Screen(
-            bottom=libqtile.bar.Bar(
-                [
-                    libqtile.widget.GroupBox(),
-                ],
-                20,
-            ),
-        ),
-        libqtile.config.Screen(
-            bottom=libqtile.bar.Bar(
-                [
-                    libqtile.widget.GroupBox(),
-                ],
-                20,
-            ),
-        ),
-    ]
-
-
-@pytest.mark.parametrize(
-    "manager",
-    [CursorWarpConfig],
-    indirect=True,
-)
-def test_cursor_warp(manager):
-    conn = xcbq.Connection(manager.display)
-    root = conn.default_screen.root.wid
-
-    assert manager.c.screen.info()["index"] == 0
-
-    manager.test_window("one")
-    manager.c.window.set_position_floating(50, 50)
-    manager.c.window.set_size_floating(50, 50)
-
-    manager.c.to_screen(1)
-    assert manager.c.screen.info()["index"] == 1
-
-    p = conn.conn.core.QueryPointer(root).reply()
-    # Here pointer should warp to the second screen as there are no windows
-    # there.
-    assert p.root_x == conftest.WIDTH + conftest.SECOND_WIDTH // 2
-    # Reduce the bar height from the screen height.
-    assert p.root_y == (conftest.SECOND_HEIGHT - 20) // 2
-
-    manager.c.to_screen(0)
-    assert manager.c.window.info()["name"] == "one"
-
-    p = conn.conn.core.QueryPointer(manager.c.window.info()["id"]).reply()
-
-    # Here pointer should warp to the window.
-    assert p.win_x == 25
-    assert p.win_y == 25
-    assert p.same_screen
 
 
 def test_switch_groups_cursor_warp(manager_nospawn):
