+++ test/test_window.py
@@ -46,181 +46,3 @@ def test_no_size_hint(manager):
     manager.c.window.set_size_floating(200, 200)
     assert manager.c.window.info()['width'] == 200
     assert manager.c.window.info()['height'] == 200
-
-
-@bare_config
-def test_min_size_hint(manager):
-    w = None
-    conn = xcbq.Connection(manager.display)
-
-    def size_hints():
-        nonlocal w
-        w = conn.create_window(0, 0, 100, 100)
-
-        # set the size hints
-        hints = [0] * 18
-        hints[0] = xcbq.NormalHintsFlags["PMinSize"]
-        hints[5] = hints[6] = 100
-        w.set_property("WM_NORMAL_HINTS", hints, type="WM_SIZE_HINTS", format=32)
-        w.map()
-        conn.conn.flush()
-
-    try:
-        manager.create_window(size_hints)
-        manager.c.window.enable_floating()
-        print(w.get_wm_normal_hints())
-        assert manager.c.window.info()['width'] == 100
-        assert manager.c.window.info()['height'] == 100
-
-        manager.c.window.set_size_floating(50, 50)
-        assert manager.c.window.info()['width'] == 100
-        assert manager.c.window.info()['height'] == 100
-
-        manager.c.window.set_size_floating(200, 200)
-        assert manager.c.window.info()['width'] == 200
-        assert manager.c.window.info()['height'] == 200
-    finally:
-        w.kill_client()
-        conn.finalize()
-
-
-@bare_config
-def test_min_size_hint_no_flag(manager):
-    w = None
-    conn = xcbq.Connection(manager.display)
-
-    def size_hints():
-        nonlocal w
-        w = conn.create_window(0, 0, 100, 100)
-
-        # set the size hints
-        hints = [0] * 18
-        hints[5] = hints[6] = 100
-        w.set_property("WM_NORMAL_HINTS", hints, type="WM_SIZE_HINTS", format=32)
-        w.map()
-        conn.conn.flush()
-
-    try:
-        manager.create_window(size_hints)
-        manager.c.window.enable_floating()
-        print(w.get_wm_normal_hints())
-        assert manager.c.window.info()['width'] == 100
-        assert manager.c.window.info()['height'] == 100
-
-        manager.c.window.set_size_floating(50, 50)
-        assert manager.c.window.info()['width'] == 50
-        assert manager.c.window.info()['height'] == 50
-
-        manager.c.window.set_size_floating(200, 200)
-        assert manager.c.window.info()['width'] == 200
-        assert manager.c.window.info()['height'] == 200
-    finally:
-        w.kill_client()
-        conn.finalize()
-
-
-@bare_config
-def test_max_size_hint(manager):
-    w = None
-    conn = xcbq.Connection(manager.display)
-
-    def size_hints():
-        nonlocal w
-        w = conn.create_window(0, 0, 100, 100)
-
-        # set the size hints
-        hints = [0] * 18
-        hints[0] = xcbq.NormalHintsFlags["PMaxSize"]
-        hints[7] = hints[8] = 100
-        w.set_property("WM_NORMAL_HINTS", hints, type="WM_SIZE_HINTS", format=32)
-        w.map()
-        conn.conn.flush()
-
-    try:
-        manager.create_window(size_hints)
-        manager.c.window.enable_floating()
-        print(w.get_wm_normal_hints())
-        assert manager.c.window.info()['width'] == 100
-        assert manager.c.window.info()['height'] == 100
-
-        manager.c.window.set_size_floating(50, 50)
-        assert manager.c.window.info()['width'] == 50
-        assert manager.c.window.info()['height'] == 50
-
-        manager.c.window.set_size_floating(200, 200)
-        assert manager.c.window.info()['width'] == 100
-        assert manager.c.window.info()['height'] == 100
-    finally:
-        w.kill_client()
-        conn.finalize()
-
-
-@bare_config
-def test_max_size_hint_no_flag(manager):
-    w = None
-    conn = xcbq.Connection(manager.display)
-
-    def size_hints():
-        nonlocal w
-        w = conn.create_window(0, 0, 100, 100)
-
-        # set the size hints
-        hints = [0] * 18
-        hints[7] = hints[8] = 100
-        w.set_property("WM_NORMAL_HINTS", hints, type="WM_SIZE_HINTS", format=32)
-        w.map()
-        conn.conn.flush()
-
-    try:
-        manager.create_window(size_hints)
-        manager.c.window.enable_floating()
-        print(w.get_wm_normal_hints())
-        assert manager.c.window.info()['width'] == 100
-        assert manager.c.window.info()['height'] == 100
-
-        manager.c.window.set_size_floating(50, 50)
-        assert manager.c.window.info()['width'] == 50
-        assert manager.c.window.info()['height'] == 50
-
-        manager.c.window.set_size_floating(200, 200)
-        assert manager.c.window.info()['width'] == 200
-        assert manager.c.window.info()['height'] == 200
-    finally:
-        w.kill_client()
-        conn.finalize()
-
-
-@bare_config
-def test_both_size_hints(manager):
-    w = None
-    conn = xcbq.Connection(manager.display)
-
-    def size_hints():
-        nonlocal w
-        w = conn.create_window(0, 0, 100, 100)
-
-        # set the size hints
-        hints = [0] * 18
-        hints[0] = xcbq.NormalHintsFlags["PMinSize"] | xcbq.NormalHintsFlags["PMaxSize"]
-        hints[5] = hints[6] = hints[7] = hints[8] = 100
-        w.set_property("WM_NORMAL_HINTS", hints, type="WM_SIZE_HINTS", format=32)
-        w.map()
-        conn.conn.flush()
-
-    try:
-        manager.create_window(size_hints)
-        manager.c.window.enable_floating()
-        print(w.get_wm_normal_hints())
-        assert manager.c.window.info()['width'] == 100
-        assert manager.c.window.info()['height'] == 100
-
-        manager.c.window.set_size_floating(50, 50)
-        assert manager.c.window.info()['width'] == 100
-        assert manager.c.window.info()['height'] == 100
-
-        manager.c.window.set_size_floating(200, 200)
-        assert manager.c.window.info()['width'] == 100
-        assert manager.c.window.info()['height'] == 100
-    finally:
-        w.kill_client()
-        conn.finalize()
